UI: Rename pluginEndpoints to agentPluginEndpoints
