+++ monkey/monkey_island/cc/next_ui/src/redux/features/api/agentPlugins/types.ts
